Give ObjectManager the ability to disable writing to old-format json project directories

+++ source/org/miradi/project/ObjectManager.java
@@ -349,8 +349,11 @@ public class ObjectManager
 		if(pool == null)
 			throw new RuntimeException("No pool for " + objectType);
 		BaseObject created = pool.createObject(this, objectId);
-		getDatabase().writeObject(created);
-		getDatabase().writeObjectManifest(getFileName(), objectType, createManifest(pool));
+		if(writeToOldProjectDirectories)
+		{
+			getDatabase().writeObject(created);
+			getDatabase().writeObjectManifest(getFileName(), objectType, createManifest(pool));
+		}
 		createdId = created.getId();
 		return createdId;
 	}
@@ -370,15 +373,19 @@ public class ObjectManager
 		if(pool.findObject(objectId) == null)
 			throw new RuntimeException("Attempted to delete missing object: " + objectType + ":" + objectId);
 		pool.remove(objectId);
-		getDatabase().writeObjectManifest(getFileName(), objectType, createManifest(pool));
-		getDatabase().deleteObject(objectType, objectId);
+		if(writeToOldProjectDirectories)
+		{
+			getDatabase().writeObjectManifest(getFileName(), objectType, createManifest(pool));
+			getDatabase().deleteObject(objectType, objectId);
+		}
 	}
 	
 	public void setObjectData(ORef objectRef, String fieldTag, String dataValue) throws Exception
 	{
 		BaseObject object = findObject(objectRef);
 		object.setData(fieldTag, dataValue);
-		getDatabase().writeObject(object);
+		if(writeToOldProjectDirectories)
+			getDatabase().writeObject(object);
 	}
 	
 	public BaseObject findObject(ORef ref)
@@ -617,4 +624,6 @@ public class ObjectManager
 	private ChainWalker diagramChainWalker;
 	private HashMap<Integer, BaseObjectInformation> pools;
 	private HashMap<ORef, ORefSet> referrerCache;
+	
+	public static boolean writeToOldProjectDirectories = true;
 }
